Changed name to DefaultChannelListener.
